get_github_input/requirements.txt
anthropic>=0.21.3
black>=24.3.0
coverage>=7.4.4
datasets>=2.18.0
docker>=7.0.0
fire>=0.6.0
flask>=3.0.2
gitpython>=3.1.43
ipykernel>=6.29.4
ipywidgets>=8.1.2
isort>=5.13.2
jsonlines>=4.0.0
matplotlib>=3.8.3
mypy>=1.9.0
networkx>=3.3
nltk>=3.8.1
openai>=1.16.0
openpyxl>=3.1.2
pebble>=5.0.7
pycg>=0.0.8
pyright>=1.1.357
pytest>=8.1.1
rich>=13.7.1
rpyc>=6.0.0
seaborn>=0.13.2
setuptools>=69.2.0
skypilot[gcp]>=0.5.0
termcolor>=2.4.0
tiktoken>=0.6.0
tomlkit>=0.12.4
tqdm>=4.66.2
click>=8.1.7
PyGithub